+++ tests/src/Functional/DeveloperAppCredentialEventTest.php
@@ -28,15 +28,14 @@ use Drupal\apigee_edge\Entity\DeveloperApp;
 use Drupal\apigee_edge\Event\AppCredentialCreateEvent;
 use Drupal\apigee_edge\Event\AppCredentialDeleteEvent;
 use Drupal\apigee_edge_test_app_keys\EventSubscriber\CreateDeleteAppKey;
-use Drupal\Core\Entity\EntityInterface;
 
 /**
- * Create, delete, update Developer App entity tests.
+ * Developer app keys, credential event test.
  *
  * @group apigee_edge
  * @group apigee_edge_developer_app
  */
-class DeveloperAppTest extends ApigeeEdgeFunctionalTestBase {
+class DeveloperAppCredentialEventTest extends ApigeeEdgeFunctionalTestBase {
 
   /**
    * The Drupal user that belongs to the developer app's developer.
@@ -48,9 +47,23 @@ class DeveloperAppTest extends ApigeeEdgeFunctionalTestBase {
   /**
    * The owner of the developer app.
    *
-   * @var \Drupal\apigee_edge\Entity\Developer
+   * @var \Drupal\apigee_edge\Entity\DeveloperInterface
    */
   protected $developer;
+
+  /**
+   * Developer app to test.
+   *
+   * @var \Drupal\apigee_edge\Entity\DeveloperAppInterface
+   */
+  protected $developerApp;
+
+  /**
+   * API product to test.
+   *
+   * @var \Drupal\apigee_edge\Entity\ApiProductInterface
+   */
+  protected $apiProduct;
 
   /**
    * {@inheritdoc}
@@ -58,44 +71,53 @@ class DeveloperAppTest extends ApigeeEdgeFunctionalTestBase {
   protected function setUp() {
     parent::setUp();
 
+    $this->installExtraModules(['apigee_edge_test_app_keys']);
+
     $this->account = $this->createAccount();
     $this->developer = Developer::load($this->account->getEmail());
+
+    $this->developerApp = DeveloperApp::create([
+      'name' => $this->randomMachineName(),
+      'status' => App::STATUS_APPROVED,
+      'developerId' => $this->developer->getDeveloperId(),
+    ]);
+    $this->developerApp->setOwner($this->account);
+    $this->developerApp->save();
+
+    $this->apiProduct = ApiProduct::create([
+      'name' => $this->randomMachineName(),
+      'displayName' => $this->randomMachineName(),
+      'approvalType' => ApiProduct::APPROVAL_TYPE_AUTO,
+    ]);
+    $this->apiProduct->save();
   }
 
   /**
    * {@inheritdoc}
    */
   protected function tearDown() {
-    $this->developer->delete();
-
+    try {
+      if ($this->developer !== NULL) {
+        $this->developer->delete();
+      }
+    }
+    catch (\Exception $exception) {
+      $this->logException($exception);
+    }
+    try {
+      if ($this->apiProduct !== NULL) {
+        $this->apiProduct->delete();
+      }
+    }
+    catch (\Exception $exception) {
+      $this->logException($exception);
+    }
     parent::tearDown();
   }
 
   /**
-   * Tests developer app entity.
+   * Developer app keys, credential event test.
    */
-  public function testCrud() {
-    /** @var \Drupal\apigee_edge\Entity\DeveloperApp $app */
-    $app = $this->createApp();
-
-    $this->assertNotEmpty($app->getAppId());
-
-    $this->assertNotEmpty(DeveloperApp::load($app->id()));
-
-    $applist = DeveloperApp::loadMultiple();
-    $this->assertContains($app->id(), array_keys($applist));
-
-    $value = $this->randomMachineName();
-    $app->setAttribute('test', $value);
-    $app->save();
-
-    /** @var \Drupal\apigee_edge\Entity\DeveloperApp $loadedApp */
-    $loadedApp = DeveloperApp::load($app->id());
-    $this->assertEquals($value, $loadedApp->getAttributeValue('test'));
-
-    $app->delete();
-  }
-
   public function testAppCredentialEvents() {
     $validateCredential = function (AppCredentialInterface $credential) {
       $prefix = apigee_edge_test_app_keys_get_prefix();
@@ -103,67 +125,35 @@ class DeveloperAppTest extends ApigeeEdgeFunctionalTestBase {
       $this->assertStringStartsWith($prefix, $credential->getConsumerSecret());
     };
 
-    $this->installExtraModules(['apigee_edge_test_app_keys']);
-    /** @var \Drupal\apigee_edge\Entity\DeveloperAppInterface $app */
-    $app = $this->createApp();
-    $credentials = $app->getCredentials();
+    $credentials = $this->developerApp->getCredentials();
     $credential = reset($credentials);
     $validateCredential($credential);
-    $app->delete();
 
     // Override (default) app key when an app is created.
-    $app = $this->createApp();
-    $apiproduct = ApiProduct::create([
-      'name' => $this->randomMachineName(),
-      'displayName' => $this->randomMachineName(),
-      'approvalType' => ApiProduct::APPROVAL_TYPE_AUTO,
-    ]);
-    $apiproduct->save();
-    $credentials = $app->getCredentials();
+    $credentials = $this->developerApp->getCredentials();
     $credential = reset($credentials);
-    /** @var \Drupal\apigee_edge\SDKConnectorInterface $connector */
     $connector = $this->container->get('apigee_edge.sdk_connector');
-    $dacc = new DeveloperAppCredentialController($connector->getOrganization(), $app->getDeveloperId(), $app->getName(), $connector->getClient());
+    $dacc = new DeveloperAppCredentialController($connector->getOrganization(), $this->developerApp->getDeveloperId(), $this->developerApp->getName(), $connector->getClient());
     $dacc->delete($credential->getConsumerKey());
 
     // Override app key on generate.
-    $dacc->generate([$apiproduct->id()], $app->getAttributes(), (string) $app->getCallbackUrl(), $app->getScopes(), 60 * 60 * 1000);
+    $dacc->generate([$this->apiProduct->id()], $this->developerApp->getAttributes(), (string) $this->developerApp->getCallbackUrl(), $this->developerApp->getScopes(), 60 * 60 * 1000);
     // Also test that related caches got invalidated, this is the reason why
     // we retrieve the credentials from the app instead of use the return
     // value of the function above.
-    $credentials = $app->getCredentials();
+    $credentials = $this->developerApp->getCredentials();
     $credential = reset($credentials);
     $validateCredential($credential);
 
     // Delete app key event.
-    /** @var \Drupal\Core\State\StateInterface $states */
     $state = $this->container->get('state');
     $dacc->delete($credential->id());
-    $this->assertNotNull($state->get(CreateDeleteAppKey::generateStateKey(AppCredentialDeleteEvent::APP_TYPE_DEVELOPER, $app->getDeveloperId(), $app->getName(), $credential->id())));
+    $this->assertNotNull($state->get(CreateDeleteAppKey::generateStateKey(AppCredentialDeleteEvent::APP_TYPE_DEVELOPER, $this->developerApp->getDeveloperId(), $this->developerApp->getName(), $credential->id())));
 
     // Create (additional) app key event.
     $credential_key = $this->randomMachineName();
     $dacc->create($credential_key, $this->randomMachineName());
-    $this->assertNotNull($state->get(CreateDeleteAppKey::generateStateKey(AppCredentialCreateEvent::APP_TYPE_DEVELOPER, $app->getDeveloperId(), $app->getName(), $credential->id())));
-
-    $app->delete();
-    $apiproduct->delete();
-  }
-
-  /**
-   * Creates new developer app.
-   *
-   * @throws \Drupal\Core\Entity\EntityStorageException
-   */
-  protected function createApp(): EntityInterface {
-    $app = DeveloperApp::create([
-      'name' => $this->randomMachineName(),
-      'status' => App::STATUS_APPROVED,
-      'developerId' => $this->developer->getDeveloperId(),
-    ]);
-    $app->setOwner($this->account);
-    $app->save();
-    return $app;
+    $this->assertNotNull($state->get(CreateDeleteAppKey::generateStateKey(AppCredentialCreateEvent::APP_TYPE_DEVELOPER, $this->developerApp->getDeveloperId(), $this->developerApp->getName(), $credential->id())));
   }
 
 }
